fixing circular dependencies
