+++ src/range/CompositeRange1D.ts
@@ -2,7 +2,7 @@
  * Created by Samuel Gratzl on 27.12.2016.
  */
 
-import Range1D from './Range1D';
+import Range1D, {ICompositeRange1D} from './Range1D';
 import Range1DGroup from './Range1DGroup';
 
 function toBase(groups: Range1DGroup[]) {
@@ -21,7 +21,7 @@ function toBase(groups: Range1DGroup[]) {
 }
 
 
-export default class CompositeRange1D extends Range1D {
+export default class CompositeRange1D extends Range1D implements ICompositeRange1D {
   constructor(public readonly name: string, public readonly groups: Range1DGroup[], base?: Range1D) {
     super(base ? base : toBase(groups));
   }
@@ -63,5 +63,9 @@ export default class CompositeRange1D extends Range1D {
 
   toString() {
     return '"' + this.name + '"{' + this.groups.join(',') + '}';
+  }
+
+  fromLikeComposite(groups: Range1DGroup[]) {
+    return new CompositeRange1D(this.name, groups);
   }
 }
