+++ src/table/ATable.ts
@@ -3,11 +3,10 @@
  */
 
 import {IPersistable} from '../index';
-import { all, parse, RangeLike} from '../range';
+import {Range, all, parse, RangeLike} from '../range';
 import {SelectAble, resolve as idtypes_resolve, IDType} from '../idtype';
 import {IVector} from '../vector';
 import {ITable, IQueryArgs} from './ITable';
-import TableView from './internal/TableView';
 import MultiTableVector from './internal/MultiTableVector';
 import {IValueType, IValueTypeDesc} from '../datatype';
 
@@ -37,6 +36,8 @@ export abstract class ATable extends SelectAble {
     return new TableView(this.root, parse(range));
   }
 
+  abstract colData(column: string, range?: RangeLike);
+
   abstract queryView(name: string, args: IQueryArgs): ITable;
 
   idView(idRange: RangeLike = all()): Promise<ITable> {
@@ -61,3 +62,100 @@ export abstract class ATable extends SelectAble {
 }
 
 export default ATable;
+
+// circular dependency thus not extractable
+/**
+ * view on the vector restricted by a range
+ * @param root underlying matrix
+ * @param range range selection
+ * @param t optional its transposed version
+ * @constructor
+ */
+export class TableView extends ATable implements ITable {
+  constructor(root: ITable, private range: Range) {
+    super(root);
+    this.range = range;
+  }
+
+  get desc() {
+    return this.root.desc;
+  }
+
+  persist() {
+    return {
+      root: this.root.persist(),
+      range: this.range.toString()
+    };
+  }
+
+  restore(persisted: any) {
+    let r: ITable = this;
+    if (persisted && persisted.range) { //some view onto it
+      r = r.view(parse(persisted.range));
+    }
+    return r;
+  }
+
+  size() {
+    return this.range.size(this.root.dim);
+  }
+
+  at(i: number, j: number) {
+    let inverted = this.range.invert([i, j], this.root.dim);
+    return this.root.at(inverted[0], inverted[1]);
+  }
+
+  col(i: number) {
+    let inverted = this.range.invert([0, i], this.root.dim);
+    return this.root.col(inverted[1]);
+  }
+
+  cols(range: RangeLike = all()) {
+    return this.root.cols(this.range.swap().preMultiply(parse(range), this.root.dim));
+  }
+
+  data(range: RangeLike = all()) {
+    return this.root.data(this.range.preMultiply(parse(range), this.root.dim));
+  }
+
+  colData(column: string, range: RangeLike = all()) {
+    return this.root.colData(column, this.range.preMultiply(parse(range), this.root.dim));
+  }
+
+  objects(range: RangeLike = all()) {
+    return this.root.objects(this.range.preMultiply(parse(range), this.root.dim));
+  }
+
+  rows(range: RangeLike = all()) {
+    return this.root.rows(this.range.preMultiply(parse(range), this.root.dim));
+  }
+
+  rowIds(range: RangeLike = all()) {
+    return this.root.rowIds(this.range.preMultiply(parse(range), this.root.dim));
+  }
+
+  ids(range: RangeLike = all()) {
+    return this.rowIds(range);
+  }
+
+  view(range: RangeLike = all()) {
+    const r = parse(range);
+    if (r.isAll) {
+      return this;
+    }
+    return new TableView(this.root, this.range.preMultiply(r, this.dim));
+  }
+
+  get idtype() {
+    return this.root.idtype;
+  }
+
+  get idtypes() {
+    return [this.idtype];
+  }
+
+  queryView(name: string, args: any): ITable {
+    throw new Error('not implemented');
+  }
+}
+
