+++ src/vector/AVector.ts
@@ -7,7 +7,8 @@
  * Created by Samuel Gratzl on 04.08.2014.
  */
 
-import {all, Range, RangeLike, range, CompositeRange1D, asUngrouped, composite, parse} from '../range';
+import {all, list as rlist, Range, RangeLike, range, CompositeRange1D, asUngrouped, composite, parse} from '../range';
+import {argSort, argFilter} from '../index';
 import {SelectAble} from '../idtype';
 import {
   categorical2partitioning,
@@ -23,7 +24,6 @@ import {
 import {computeStats, IStatistics, IHistogram, categoricalHist, hist} from '../math';
 import {IVector} from './IVector';
 import {IStratification} from '../stratification';
-import VectorView from './internal/VectorView';
 import StratificationVector from './internal/StratificationVector';
 /**
  * base class for different Vector implementations, views, transposed,...
@@ -155,3 +155,88 @@ export abstract class AVector extends SelectAble {
 }
 
 export default AVector;
+
+
+/**
+ * view on the vector restricted by a range
+ * @param root underlying matrix
+ * @param range range selection
+ * @param t optional its transposed version
+ * @constructor
+ */
+export class VectorView extends AVector {
+  constructor(root: IVector, private range: Range) {
+    super(root);
+  }
+
+  get desc() {
+    return this.root.desc;
+  }
+
+  persist() {
+    return {
+      root: this.root.persist(),
+      range: this.range.toString()
+    };
+  }
+
+  size() {
+    return this.range.size(this.root.dim)[0];
+  }
+
+  at(i: number) {
+    const inverted = this.range.invert([i], this.root.dim);
+    return this.root.at(inverted[0]);
+  }
+
+  data(range: RangeLike = all()) {
+    return this.root.data(this.range.preMultiply(parse(range), this.root.dim));
+  }
+
+  names(range: RangeLike = all()) {
+    return this.root.names(this.range.preMultiply(parse(range), this.root.dim));
+  }
+
+  ids(range: RangeLike = all()) {
+    return this.root.ids(this.range.preMultiply(parse(range), this.root.dim));
+  }
+
+  view(range: RangeLike = all()) {
+    const r = parse(range);
+    if (r.isAll) {
+      return this;
+    }
+    return new VectorView(this.root, this.range.preMultiply(r, this.dim));
+  }
+
+  get valuetype() {
+    return this.root.valuetype;
+  }
+
+  get idtype() {
+    return this.root.idtype;
+  }
+
+  get idtypes() {
+    return [this.idtype];
+  }
+
+  /*get indices() {
+   return this.range;
+   }*/
+
+  sort(compareFn?: (a: IValueType, b: IValueType) => number, thisArg?: any): Promise<IVector> {
+    return this.data().then((d) => {
+      const indices = argSort(d, compareFn, thisArg);
+      return this.view(this.range.preMultiply(rlist(indices)));
+    });
+  }
+
+  filter(callbackfn: (value: IValueType, index: number) => boolean, thisArg?: any): Promise<IVector> {
+    return this.data().then((d) => {
+      const indices = argFilter(d, callbackfn, thisArg);
+      return this.view(this.range.preMultiply(rlist(indices)));
+    });
+  }
+}
+
